misc casts in bignum, ecp

+++ library/bignum_core.c
@@ -58,7 +58,7 @@ size_t mbedtls_mpi_core_bitlen(const mbedtls_mpi_uint *A, size_t A_limbs)
     for (i = ((int) A_limbs) - 1; i >= 0; i--) {
         if (A[i] != 0) {
             j = biL - mbedtls_mpi_core_clz(A[i]);
-            return (i * biL) + j;
+            return ((unsigned int) i * biL) + j;
         }
     }
 
@@ -585,7 +585,7 @@ int mbedtls_mpi_core_get_mont_r2_unsafe(mbedtls_mpi *X,
     int ret = MBEDTLS_ERR_ERROR_CORRUPTION_DETECTED;
 
     MBEDTLS_MPI_CHK(mbedtls_mpi_lset(X, 1));
-    MBEDTLS_MPI_CHK(mbedtls_mpi_shift_l(X, N->n * 2 * biL));
+    MBEDTLS_MPI_CHK(mbedtls_mpi_shift_l(X, N->n * 2u * biL));
     MBEDTLS_MPI_CHK(mbedtls_mpi_mod_mpi(X, X, N));
     MBEDTLS_MPI_CHK(mbedtls_mpi_shrink(X, N->n));
